+++ packages/crypto-key-pair-bls12-381/source/service-provider.ts
@@ -1,0 +1,52 @@
+import { Identifiers } from "@mainsail/constants";
+import { injectable, Selectors } from "@mainsail/container";
+import type { Contracts } from "@mainsail/contracts";
+import { Providers } from "@mainsail/kernel";
+
+import { KeyPairFactory } from "./pair.js";
+import { PrivateKeyFactory } from "./private.js";
+import { PublicKeyFactory } from "./public.js";
+import { schemas } from "./schemas.js";
+import { PublicKeySerializer } from "./serializer.js";
+
+@injectable()
+export class ServiceProvider extends Providers.ServiceProvider {
+	public async register(): Promise<void> {
+		this.app
+			.bind(Identifiers.Cryptography.Identity.PublicKey.Size)
+			.toConstantValue(48)
+			.when(Selectors.anyAncestorOrTargetTagged("type", "wallet"));
+
+		this.app
+			.bind(Identifiers.Cryptography.Identity.KeyPair.Factory)
+			.to(KeyPairFactory)
+			.inSingletonScope()
+			.when(Selectors.anyAncestorOrTargetTagged("type", "wallet"));
+
+		this.app
+			.bind(Identifiers.Cryptography.Identity.PrivateKey.Factory)
+			.to(PrivateKeyFactory)
+			.inSingletonScope()
+			.when(Selectors.anyAncestorOrTargetTagged("type", "wallet"));
+
+		this.app
+			.bind(Identifiers.Cryptography.Identity.PublicKey.Factory)
+			.to(PublicKeyFactory)
+			.inSingletonScope()
+			.when(Selectors.anyAncestorOrTargetTagged("type", "wallet"));
+
+		this.app
+			.bind(Identifiers.Cryptography.Identity.PublicKey.Serializer)
+			.to(PublicKeySerializer)
+			.inSingletonScope()
+			.when(Selectors.anyAncestorOrTargetTagged("type", "wallet"));
+
+		this.#registerSchemas();
+	}
+
+	#registerSchemas(): void {
+		for (const schema of Object.values(schemas)) {
+			this.app.get<Contracts.Crypto.Validator>(Identifiers.Cryptography.Validator).addSchema(schema);
+		}
+	}
+}
